fixed windows tcp connection issues
