+++ questdb-rs/src/ingress/tls.rs
@@ -90,7 +90,6 @@ fn unpack_os_native_certs(
 
 #[cfg(feature = "tls-native-certs")]
 fn add_os_roots(root_store: &mut RootCertStore) -> crate::Result<()> {
-    eprintln!("add_os_roots :: (1)");
     let os_certs = unpack_os_native_certs(rustls_native_certs::load_native_certs())?;
 
     let (valid_count, invalid_count) = root_store.add_parsable_certificates(os_certs);
@@ -101,7 +100,6 @@ fn add_os_roots(root_store: &mut RootCertStore) -> crate::Result<()> {
             invalid_count
         ));
     }
-    eprintln!("add_os_roots :: (2) valid_count: {valid_count}, invalid_count: {invalid_count}");
     Ok(())
 }
 
@@ -131,7 +129,6 @@ impl TlsSettings {
         ca: CertificateAuthority,
         roots: Option<&Path>,
     ) -> Result<Option<Self>> {
-        eprintln!("TlsSettings::build :: (1) enabled: {enabled}, ca: {ca:?}, roots: {roots:?}");
         if !enabled {
             return Ok(None);
         }
