+++ system_test/tls_proxy/Cargo.toml
@@ -6,10 +6,10 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tokio = { version = "1.32.0", features = ["full"] }
+tokio = { version = "1.46.0", features = ["full"] }
 tokio-rustls = "0.26.2"
 rustls = "0.23.28"
-rustls-pemfile = "2.0.0"
-argh = "0.1.12"
-anyhow = "1.0.75"
-futures = "0.3.29"
+rustls-pemfile = "2.2.0"
+argh = "0.1.13"
+anyhow = "1.0.98"
+futures = "0.3.31"
